Fixed Try Catch to properly catch errors
Removed the command that cleared the $error log. This was a left-over from testing.

Added Try, Catch to removing the old Spicetify folders from the user's home directory.

Fixed the old error catcher to actually work and to use a modern up-to-date function.

Wrapped the end of the script in a function so that it can be more easily troubleshooted and consistent.